requirements.txt
fastapi>=0.116.1
uvicorn>=0.35.0
numpy>=2.0.0
scikit-learn>=1.3.0
Pillow>=10.0.0
ultralytics>=8.0.0